+++ src/Elements/Ul.ts
@@ -1,0 +1,9 @@
+import CompositeElement from './CompositeElement';
+
+class Ul extends CompositeElement<HTMLUListElement> {
+  protected renderElement(): void {
+    this.rendered = document.createElement('ul');
+  }
+}
+
+export default Ul;