+++ src/Interfaces/MenuDictionary.ts
@@ -1,0 +1,6 @@
+interface MenuDictionary {
+  title: string;
+  items: { [item: string]: number };
+}
+
+export default MenuDictionary;