+++ src/Interfaces/Renderable.ts
@@ -1,0 +1,5 @@
+interface Renderable<T extends Node> {
+  render(): T;
+}
+
+export default Renderable;